chore: Use KeyValueStoreManager instead of etcd::Client (#3822)

+++ lib/bindings/c/src/lib.rs
@@ -154,7 +154,7 @@ fn dynamo_create_kv_publisher(
     {
         Ok(drt) => {
             let backend = drt.namespace(namespace)?.component(component)?;
-            KvEventPublisher::new(backend, worker_id, kv_block_size, None)
+            KvEventPublisher::new(backend, worker_id as u64, kv_block_size, None)
         }
         Err(e) => Err(e),
     }
