+++ lib/llm/tests/block_manager.rs
@@ -216,7 +216,7 @@ pub mod llm_kvbm {
     impl DynamoKvbmRuntimeConfigBuilder {
         pub fn build(self) -> Result<kvbm::config::KvManagerRuntimeConfig> {
             let (runtime, nixl) = self.build_internal()?.dissolve();
-            let worker_id = runtime.primary_lease().unwrap().id() as u64;
+            let worker_id = runtime.primary_lease().unwrap().id();
             Ok(kvbm::config::KvManagerRuntimeConfig::builder()
                 .worker_id(worker_id)
                 .cancellation_token(runtime.primary_token().child_token())
@@ -247,7 +247,7 @@ pub mod llm_kvbm {
     impl DynamoEventManager {
         pub fn new(component: Arc<KVBMDynamoRuntimeComponent>) -> Self {
             let (tx, rx) = mpsc::unbounded_channel();
-            let worker_id = component.drt().primary_lease().unwrap().id() as u64;
+            let worker_id = component.drt().primary_lease().unwrap().id();
             component.drt().runtime().secondary().spawn(async move {
                 worker_task(component, rx).await;
             });
@@ -296,7 +296,7 @@ pub mod llm_kvbm {
                         event_id: event_id_counter,
                         dp_rank: 0,
                     };
-                    let router_event = RouterEvent::new(worker_identifier as i64, event);
+                    let router_event = RouterEvent::new(worker_identifier, event);
                     event_id_counter += 1;
                     if let Err(e) = component_clone
                         .batch_tx
@@ -316,7 +316,7 @@ pub mod llm_kvbm {
                         event_id: event_id_counter,
                         dp_rank: 0,
                     };
-                    let router_event = RouterEvent::new(worker_identifier as i64, event);
+                    let router_event = RouterEvent::new(worker_identifier, event);
                     event_id_counter += 1;
                     if let Err(e) = component_clone
                         .batch_tx
